Added js and html to new folder for bundlin and removed from static
